requirements.txt
boto3==1.9.50
botocore==1.12.50
docutils==0.14
jmespath==0.9.3
psycopg2-binary==2.7.6.1
python-dateutil==2.7.5
python-dotenv==0.9.1
s3transfer==0.1.13
six==1.11.0
urllib3==1.24.1
csvdiff==0.3.3
